refactor(leads): extract LeadStatus + STATUS_STYLES to leads/constants (#72)
Invert the direction of dependency between components/leads/ and
components/inbox/. Before: LeadCard imported STATUS_STYLES and
InboxLead['status'] from inbox/constants, which made leads depend on
inbox. After: leads/constants owns the LeadStatus type and STATUS_STYLES
map; inbox/constants re-exports them for backward compatibility and
uses LeadStatus inside its own InboxLead interface.

  frontend/src/components/leads/constants.ts   (new, owns the primitives)
  frontend/src/components/inbox/constants.ts   (re-exports + inbox-specific)
  frontend/src/components/leads/LeadCard.tsx   (imports from leads/)
  frontend/src/components/leads/LeadCard.test.tsx

Pure structural refactor — no behavioural delta. All 367 frontend tests
green, lint + build clean.

Acceptance from #72:
- grep -r 'from "@/components/inbox' src/components/leads/ → empty ✓
- All tests pass ✓
- Lint + build clean ✓

Closes #72.
